requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
prisma==0.10.1
passlib==1.7.4
python-jose==3.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9 